+++ src/globals/res.lua
@@ -5,11 +5,6 @@ return {
     TILESET = "res/img/tileset.png",
     PARTICLE = "res/img/rectangle.png",
 
-    -- Shaders
-    WAVE_SHADER = "res/shaders/wave.glsl",
-    BLOOM_SHADER = "res/shaders/bloom.glsl",
-    GLITCH_SHADER = "res/shaders/glitch.glsl",
-
     -- Sounds
     EDIT_SOUND = "res/audio/edit_theme.mp3",
     MAIN_SOUND = "res/audio/main_theme.mp3",